optim, cgen, p1: widen binary-search indices from uint8_t to int
Three sites used `uint8_t hi, lo, mid` for indices into alphabetically
sorted token tables. When the search probes mid=0 and the comparison
says "go left", `hi = mid - 1` underflows uint8_t (wraps to 255) and
the loop blunders into out-of-bounds table entries instead of
terminating. Currently latent — with table sizes <=110 entries and
working comparisons (now also fixed in find_token via #1), normal
convergence reaches lo > hi before mid hits 0. But it is the same
class of bug as #1: an arithmetic operation that depends on signed
semantics performed on an unsigned narrow type.

Sites widened:
  optim/optim.c:1190  find_token        (high/low/mid)
  cgen/cgen.c:1028    sub_1B2           (hi/lo/mid)
  p1/lex.c:394-396    parseName         (mid/lo/hi)

Cost is zero — these are stack locals in tight loops, the compiler
keeps them in registers regardless of declared type on every modern
target.

Closes #4.

+++ cgen/cgen.c
@@ -1025,7 +1025,7 @@ int sub_13D(int a, int b) {
  * Token search		(Binary search algorithm is used)
  *********************************************************/
 char sub_1B2(register char *target) {
-    uint8_t hi, lo, mid;
+    int hi, lo, mid;
     int cmp;
 
     /* clang-format off */
